removes Rspec warnings
* Because we currently on Lita v4.7 we probably don't need to maintain
backwards-compatibility with Lita v3. This drops the Rspec warning:
```ruby
WARNING: Lita 3 compatibility mode is currently enabled. It is recommended that you disable it by adding the following code to spec/spec_helper.rb:

    require "lita/rspec"
    Lita.version_3_compatibility_mode = false

This will ensure that your test suite works with all the changes introduced in Lita 4. The compatibility mode will be removed in Lita 5.
```